+++ google2/src/test/scala/org/broadinstitute/dsde/workbench/google2/GoogleDataprocManualTest.scala
@@ -6,7 +6,7 @@ import cats.effect.concurrent.Semaphore
 import cats.effect.{Blocker, IO}
 import cats.mtl.Ask
 import com.google.cloud.compute.v1.Operation
-import com.google.cloud.dataproc.v1.{Cluster, ClusterOperationMetadata}
+import com.google.cloud.dataproc.v1.Cluster
 import io.chrisdavenport.log4cats.slf4j.Slf4jLogger
 import org.broadinstitute.dsde.workbench.model.TraceId
 import org.broadinstitute.dsde.workbench.model.google.GoogleProject
@@ -45,7 +45,7 @@ final class GoogleDataprocManualTest(pathToCredential: String,
   def callResizeCluster(cluster: String,
                         numWorkers: Option[Int],
                         numPreemptibles: Option[Int]
-  ): IO[Option[ClusterOperationMetadata]] =
+  ): IO[Option[DataprocOperation]] =
     dataprocServiceResource.use { dataprocService =>
       dataprocService.resizeCluster(project, region, DataprocClusterName(cluster), numWorkers, numPreemptibles)
     }